Configure a root disk & ephemeral disk for baremetal nodes
The current NODE_DISK variable (or disk option in testenv.json file)
specifies the size of the disk at our disposal. Add new ROOT_DISK
variable that specifies the size of the root partition on the the disk
and we then calculate and assign the remaining disk space to the
epheramal disk that will store all persistent data for baremetal nodes.

Change-Id: I1a941e1120fd4091311b7f80e145c47715ec2465