Write network configuration into seeds config.json
Updates are devtest_seed script so that the jq command to
inject information into config.json includes the latest
baremetal network info.

The existing jq command was split into a multiline string to
make it easier to edit and maintain.

As part of this change we also add the minimal required
fields for the "baremetal-network" config section to
the devtest_testenv.sh output for backwards compat
with existing users.

Change-Id: Ib9d309b89fb8f5dea8e20e154206ffcfec6ebfd7