Add some clarity to the first-time user experience
* Explicitly instruct the user to set $TRIPLEO_ROOT. Although there's a
  default value provided with no comments, this should be enough to
  prompt the user to set it to some other value if they prefer.
* Remove the aspirational instruction to "git checkout <stable-branch" -
  we can add that back just as soon as we get one.
* devtest_variables.sh tells the user to do one thing, but does
  something different. Remove the surprise by having the script do as it
  says to do.

As a side-effect, setting TRIPLEO_ROOT explicitly avoids the first-timer
trap of trying to run devtest_setup.sh in the second step of
devtest.thml - prior to this change, this was implicitly relying on
$TRIPLEO_ROOT/tripleo-incubator/scripts having devtest-setup.sh in it,
which is true on the second run but not the first.

This also avoids the trap where a user does the initial checkout into
~/src/tripleo-incubator (or ~/openstack/tripleo-incubator, or ...) and
then later edits scripts there and wonders why the changes don't get
picked up.

Change-Id: I85e43df1cf1201b07131acf4985edfdf99cef30c