Add some clarity to the first-time user experience
If a user follows the instructions in devtest, they'll first check out
triple-incubator somewhere; then run scripts inside it, which (unknown
to them) check out a different copy of tripleo-incubator into ~/.cache.

This causes several surprises for the user: they'll be surprised to find
that lots of disk images have been created, but they can't find those
images. They'll be surprised when they edit scripts inside their
original checkout, but those scripts are ignored when running the
devtest_* scripts. They may well be surprised to find that none of the
work they've done inside TRIPLEO_ROOT gets backed up, because most
backup systems will ignore data under ~/.cache.

This change moves the instructions about setting TRIPLEO_ROOT from
devtest_variables.sh to the top of devtest.sh, so it's seen by the user
before they do their initial checkout. Instructions for that checkout
are tweaked so that the checkout goes into TRIPLEO_ROOT. The user should
no longer be surprised by the location of the artifacts created by the
scripts, as they'll have explicitly configured TRIPLEO_ROOT.

As a side-effect, setting TRIPLEO_ROOT explicitly avoids the first-timer
trap of trying to run devtest_setup.sh in the second step of
devtest.html - prior to this change, this was implicitly relying on
$TRIPLEO_ROOT/tripleo-incubator/scripts having devtest-setup.sh in it,
which is true on the second run but not the first. This means we can get
rid of some undocumented code that was working around this first-time
scenario

Some other improvements have been made. Now that we have a stable
branch, the docs for checkout out a stable branch have been clarified
with a real example, and split into their own section of the docs,
labelled as being an optional step.

Change-Id: I85e43df1cf1201b07131acf4985edfdf99cef30c