Allow to use a cirros image as the end user image
For testing, there is no need to have a full $NODE_DIST image.
Using a cirros image, the memory footprint (both memory and disk) is
smaller and there is no need to create the image using dib.

By default, a NODE_DIST image with the vm element is build. This
can be overrided with the env. variable USE_CIRROS=1

Change-Id: Id0e6177068644ca2da3ccc30ca33e3c9424687eb