Add pm_type to the nodes list.
Ironic, unlike Nova-BM supports different power managers per node.

In order to configure nodes we thus need to know the desired power
manager.

This is separate to the main Ironic support patch in that it has to be
deployed on the testenv machines for CI to be able to consume it.

Change-Id: Ie4c91e039f6788b2a92909b0e0a9e540e092283c