SSH key and virtual_power_driver not used on H/W
For nova baremetal we only need the ssh details when using the
nova.virt.baremetal.virtual_power_driver.VirtualPowerManager
power driver. Hence added --sys-target which can be 'baremetal'
or 'virtual' as top level choices in devtest. This controls
placing the ssh details in the testenv or not, as well as
switching the power manager in a fully qualified manner.

Change-Id: I585a7503eeb2521a944dcfddc22c025a5bce9b3f