Generate better bm network mac addresses for VMs
Previously we relied on libvirt to generate all of
our MAC addresses in configure-vms. This change
updates configure-vm so that we generate our
own inline MACs for the baremetal network (only)
such that the 2nd MSB bit is set high. (making
it so they aren't identified as locally administered).

This effectively allows our fake baremetal VMs to more
accurately behave like real hardware and fixes issues with
bridge/DHCP configurations which rely on the fact that bridges
assume the MAC address of the lowest attached NIC.

Change-Id: I052218d88b2d4362ad5eb99576e7064aee47863f