Allow to set seed node cpus and memory
Increasing the number of cpus and the memory for the seed vm,
has impact in the time needed to boot the undercloud when the
images are built compressed  (without the -u option).
Booting times of the undercloud are reduced about 35% (More data
in the bug comments)

This can be done with the SEED_CPU and SEED_MEM env. variables.
(Defaults to 1 CPU and 2048M RAM, as current ones)

Partial-Bug: #1295732

Change-Id: Iac68701408696ae1eb598bbf7e2a3d0ad877e116