Enable selective dib configuration per image
Enable a mechanism by which dib variables can be passed selectively
into an image from the environment during a full devtest.sh run.

DIB elements often take arguments in the form of environment variables,
and there is no way to differentiate which image a variable in the user
environment should apply to when running devtest, as the same variable
would be used in each execution of diskimage-builder during the run.

This change runs diskimage-builder in a sub-shell after setting the
element variables from image-specific prefixes, thus allowing the user
to specify DIB variables on a per-image basis.

These variables prefixes are:
    * SEED_DIB_VAR_
    * UNDERCLOUD_DIB_VAR_
    * OVERCLOUD_CONTROL_DIB_VAR_
    * OVERCLOUD_COMPUTE_DIB_VAR_

So to pass DIB_NO_PYPI_PIP=1 to only the undercloud for example,
prior to running devtest.sh you would:
    export UNDERCLOUD_DIB_VAR_DIB_NO_PYPI_PIP=1

Change-Id: I42745756bbf2f5b2a79836dea61dca3a6061cc50