Make setup-clienttools optional
This patch makes it optional for the case when people want to use
packaged client tools on the host.  If they're building images
from packages, they'll likely want to use packages on the host as well.

This also allows for the case when it's desireable to have the
dependencies installed, but not the client tools.

Adds a new variable "INSTALL_CLIENTTOOLS" and defaults to true to
not disturb any current script behavior.

Change-Id: I13bc19fb7d4a9fd347eb09ef0accc50f28ce0475