Run the overcloud with an SSL enabled public IP
Enable https on the overcloud by default.

Use certs specified by SSLBASE or else generate certs automatically.

We use a virtual IP to provide a known IP address which we use as
the Common Name when automatically generating certs. (We cannot use a
dynamically allocated IP since we need to know it beforehand in order
o generate a cert with a matching Common Name).

Partially implements: blueprint tripleo-ssl-overcloud

Full set of blueprint changes:

https://review.openstack.org/#/c/85098
https://review.openstack.org/#/c/85099
https://review.openstack.org/#/c/85100

Change-Id: I68af4da3c510a14fc2ea3eb1894590cc79cf8459