Update seed element to use custom baremetal-network
in change I6e89dc98 config was added to configure custom networking for
the seed,undercloud and overcloud nodes.
Unfortunately that change did not add code to update the seed image
creation process, so some elements still get the default networking.

This change adds code to update the local.json config file in the
seed-stack-config element with the contents of the baremetal-networks
object in testenv.json

Change-Id: I6d15757ea88dff186042767f004a52b09a591e19