Remove boot-seed-vm need to know ssh key.
This has always been a bit of a layering oddness - other hypervisors
don't check ssh works. We may be able to remove the ssh glue
altogether, but this is more conservative - I just remove the need
to actually log in. An additional plus is that this is cheaper -
we're not doing a full ssh handshake.

I also noticed some plain wrong prose about what boot-seed-vm does
vis-a-vis SSH keys, so fixed that up.

Change-Id: Ic357dc882308de95f4f69e5596cfd4896ba6085a