Make bootstrapping TripleO development easier.
Folk find getting TripleO up and running tricky - they don't know to
look in ~/.cache/tripleo, or that they need to create an rc file
before doing anything to avoid that.

The approach here to tackle that is to supply a very small script that
is location independent - we can put a link to download it from gitweb
in the main TripleO docs and then its wget; run; away-we-go.

This is still technically WIP: I haven't linked it into the docs, and
the design isn't finalised - e.g. perhaps this script should accept
a destination directory - e.g. download, run configure-tripleo
/path/to/desired/tripleo/root.

Change-Id: I2fe6a03ce3be4624025dc2bc0ec36d7dcd128b4b