Rough POC of easy bootstrap for Tripleo devs
Putting up for feedback and thoughts.

The basic idea is that this can b e *outside* the tripleo devtest
docs, used as a meta tool to get folk up and running easily. As such
it doesn't need all the complex deps etc, and we can tell folk 'run it
where you want the source code'.

Change-Id: I2fe6a03ce3be4624025dc2bc0ec36d7dcd128b4b