Allow adding passwords in setup-*-passwords
By default we don't overwrite password files when running devtest,
which means that when a new password is needed the existing
password files must be deleted and regenerated or the new password
will not exist.  This change checks for the existence of each
password individually so that new passwords can simply be appended
to the password files.

Closes-Bug: #1314615

Change-Id: I746749c6cc07e7b43bd3bc327b62098d72e317cf