Make time out 60 mins as like the Heat default.
Make timeout for wating for the undercloud to compleat it create
match that of the Heat default (60 mins for the Heat stack-create).
As on slow networks with real HW the set-up time can be slow. Also
if the stack-create fails the loop will exit so this change should
be harmless.

Change-Id: Iec8aafb73f0b9da72ae9a94d9a4e7c91bd29e496