Add specific doc build option to tox
Add a doc venv to align with other OpenStack projects. Useful
for doc testing locally.

As part of this change we also update the devtest.sh docs
generation example to be a bit more generic so as not to
refer to a specific release (which may or may not
have the new tox -edocs command).

Change-Id: I37771db339d864b2b938d1c02d15d38b07093493