Make pull-tools ignore errors if USE_CACHE is set.
If I'm trying to work offline and I've set USE_CACHE, it seems odd that
pull-tools will abort a run because it can't talk to the git server.

This change makse it simply continue if a git pull fails. This is
potentially too broad (eg, it would ignore errors caused by the local
repo being corrupt, or a ff-only merge not being possible.
)

Change-Id: I9d6329567bdf7f0725624fca3b58cd9487e3d7c3