Remove obvious bashisms like [[ and ==
TripleO is targetting bash for scripts, but I think we should try and
avoid bashisms, preferring POSIX constructs if they are easy-to-use.
To that end, I'm ponying up the code. [[ and == are a good examples,
they are both bashisms, and POSIX has simple-to-use analogues.

I have also driven-by some style-matching changes in set-source-vars.

Change-Id: I573cca48776b6a4e99605170d6428608ab43d47a