Remove obvious bashisms like [[ and ==
TripleO is targetting bash for scripts, but I think we should prefer
to use POSIX if and only if it's easy for us to do so. To that end,
I'm ponying up the code. [[ and == are a good example of that, they
are both bashisms, and POSIX has simple-to-use alternatives.

I have also driven-by some style-matching changes in set-source-vars.

Change-Id: I573cca48776b6a4e99605170d6428608ab43d47a