Rebuild, but dont reconfigure undercloud services
Its useful to be able to to reconfigure the undercloud
heat stack, but only need to run init-keystone etc. if
the undercloud is being created for the first time.
This relates to Change-Id: Iaa6a11c080dd77f45b7a2e85cbb06d2769947be0
which added the ability to update the undercloud
heat stack as it is done in the overcloud.

Change-Id: I394cd028b0e258a41d39d7a7ae0319596cdc6ccf