Clean up heat env file code
There's a lot of boilerplate in this section of code which makes it a
bit hard to read and increases the chance of inconsistencies creeping
in. Also jq is called about 20 times when it only needs to be called
twice.

Use jq's + operation to merge the env vars over the saved JSON.

Change-Id: I6d7a7081e15452f1ca090610a3a068cfe30b531b