Let cleanup-env only clean a specified environment
devtest_testenv allows the user to specify an environment number. This
will:

 -- Create a seed named seed_$NUM
 -- Create a bridge named brbm$NUM
 -- Create baremetal VMs with names prefixed with baremetalbrbm$NUM

This is used on our CI environment to allow one set of physical nodes to
host several test environments.

Prior to this change, cleanup-env is very insensitive and simply cleans
up all the virsh nodes and disks that it thinks are related to devtest.

This change replicates this logic to allow cleanup-env to clean up only
one specific environment, rather than cleaning up all environments on the
machine.

This is part of a larger plan to make devtest_testenv.sh safe to re-run.
The goal is to be able to cleanup the existing environment and then
create a new environment from scratch. Once this change lands,
devtest_testenv.sh can use cleanup-env.sh to clean up just the specified
environment.

Change-Id: I1b10abda0209f1c58c6c2612542d8a6cd42628e8