Clean up the any existing environment first
This change lets devtest_testenv.sh use cleanup-env.sh's ability to
clean up just a single environment to make sure we always start from a
clean slate when creating an environment.

This removes the need for us to provide dire warnings to users about
only running devtest_testenv.sh once; and makes it simpler to run a test
with a known-clean environment.

Change-Id: I4c7ce45fd064335b73b52569cad83974c19219ee