Make disk bus type configurable
This patch makes it possible to change the bus type from the
default 'sata', for example 'scsi' to enable saving or migrating
the VM.

Both baremetal and seed node setups have been changed to use this
feature. Export LIBVIRT_DISK_BUS_TYPE in the environment with the
required bus type.

Closes-Bug: #1316675

Change-Id: I1f9961828dab1e3506a959b7880c67be8ecbebe5