Simplify creation/cleanup of testenvs
In I1b10abda0209f1c58c6c2612542d8a6cd42628e8, the way cleanup-env had to
second-guess what testenv might have done (and even then, couldn't
guarantee that it made the right choice and had to accept an explicit
override for the seed VM's name) caused a significant amount of
confusion.

Thinking about this overnight, I came to believe that the cause of this
confusion is the fact that two unrelated parameters to
devtest_testenv.sh are being entangled in the choice of name for the
seed_vm. Many people who looked at cleanup-env made the assumption that
if "-n N" is passed to _testenv.sh, the seed name will be seed_N.

This seems to me as though it's a reasonable assumption, and tweaking
the code to behave this way makes it simpler.

To the best of my knowledge, the -n and -b options were added to
_testenv to support CI, which always uses them together. This change
should make no difference to CI.

This change *may* affect people who previously used -n without -b, but
as I don't see how they'd have a working setup with >1 environments all
competing for the single VM named "seed", I don't believe this is a
problem.

Change-Id: I3a0d4315aa3e7d5281db827f01b681d7d8c3128d