Use mount-ephemeral element in undercloud and overcloud
We don't need explicitly use use-ephemeral element anymore as
this will only provide a tool to make it easier to use the
ephemeral disk. But the mount-ephemeral element will configure
images to mount this disk as early as possible so that it is
more useful.

This depends on the I002e8b5728a4edf26fb3b339f027e02871bf2304
patchset landing in tripleo-image-elements first.

Change-Id: Ibe6c0087316e93f67784183ab8157d9ca1c1a891