Provide useful error early if not enough nodes
Change Icff5e728148f2744f3fc7c6e99d744be9fda180b set the number of
nodes created for a test environment to a fixed value.

Document the reason for the change, and also error early if not enough
nodes are created, so that we don't unnecessarily waste developer time
debugging the failure of an overcloud deploy due to lack of resources.

Change-Id: I93c09e1905ca4eb6e9e580ecd0e9d4b8fcf1339e