Revert to correctly determining nodes needed
Change Icff5e728148f2744f3fc7c6e99d744be9fda180b set the number of
nodes created for a test environment to a fixed value for reasons
that were not well explained in the commit message.

This effectively broke the ability of the OVERCLOUD_*SCALE variables
to operate correctly and also created a large number of VMs on
developer workstations unnecessarily.

After discussion in IRC it was stated that a typical developer workflow
was made easier if more vms than needed were created by the testenv
script, so this change creates a minimum of 15 vms, but also honours
the pre-existing NODE_CNT variable and will create enough vms for the
settings of OVERCLOUD_*SCALE, or error.

Closes-bug: 1319473
Change-Id: I93c09e1905ca4eb6e9e580ecd0e9d4b8fcf1339e