Revert to correctly determining nodes needed
Change Icff5e728148f2744f3fc7c6e99d744be9fda180b set the number of
nodes created for a test environment to a fixed value for reasons
that were not well explained in the commit message.

This effectively broke the ability of the OVERCLOUD_*SCALE variables
to operate correctly and also created a large number of VMs on
developer workstations unnecessarily.

This sets the value back to being determined from the other variables
as is appropriate.

Closes-bug: 1319473
Change-Id: I93c09e1905ca4eb6e9e580ecd0e9d4b8fcf1339e