Revert to correctly determining nodes needed
Change Icff5e728148f2744f3fc7c6e99d744be9fda180b set the number of
nodes created for a test environment to a fixed value for reasons
that were not well explained in the commit message.

This effectively broke the ability of the OVERCLOUD_*SCALE variables
to operate correctly and also created a large number of VMs on
developer workstations unnecessarily.

This sets the value back to being determined from the other variables
as is appropriate.

It also allows overriding of the value by setting NODE_CNT in the
calling environment, as reading comments in the previous change it
seems the reason for setting it to a fixed number was for TripleoCI
purposes.  My opinion is that the TripleCI scripts are the appropriate
place to fix values, and that these user/developer-focused scripts
should remain correct for users/developers without requiring an extra
arbitrary variable to be set to match actions of environment creation
to their previously expressed desire through the use of
OVERCLOUD_*SCALE variables.

Closes-bug: 1319473
Change-Id: I93c09e1905ca4eb6e9e580ecd0e9d4b8fcf1339e