Move user's group membership setup in separate script
This change moves the code used to add current user to the
libvirt group out of install-dependencies and in an own script
launched _at the end_ of devtest_setup.

Moving it at the end of devtest_setup is useful to people
running the devtest_* scripts one by one, the process now does
not get interrupted in between the steps so there is no need
to re-run devtest_setup after relogin.

Change-Id: Ica804c24bd2cbf70a7f2b9a34fa2e0495c85c456
Partial-Bug: #1195046