Move user's group membership setup in own script
This change moves the code needed to add current user to the
libvirt group out of install-dependencies and in an own script
launched _at the end_ of devtest_setup.

Moving it at the end of devtest_setup is also helpful to people
running the devtest_* scripts one by one as the process is not
interrupted in between the steps; they don't have to re-run
devtest_setup but can instead move forward with devtest_testenv
after they relogin.

Change-Id: Ica804c24bd2cbf70a7f2b9a34fa2e0495c85c456
Partial-Bug: #1195046