Get rid of overspecification in vm.xml.
We don't need to pass the appropriate qemu binary into the vm template,
as libvirt will inject the proper one for us.  Similarly, we don't need
to tell libvirt what specific model of PC hardware to emulate, because
it will pick one that is appropriate to the devices we specify.

Change-Id: Id3762cce9b2ba9f96c1e7e79fad9a4bf8770bf7a