Enable running devtest in a Docker container.
This modifies install-dependencies to make sure that all the packages
needed are installed for opensuse and that the required device nodes and
libvirt changes are in place for the rest of devtest.sh.

It also adds a wrapper script named docker_deploy.sh that wraps the
entire devtest.sh run in a docker container.

This changeset relies on:

 * https://review.openstack.org/#/c/91867/
 * https://review.openstack.org/#/c/93912/
 * https://review.openstack.org/#/c/108375/

Change-Id: I745b87eb5b9d290d1be74dbde08aa4293cf1c2d3