Allow user-specified timeout for stack creation
The 60 min. timeout on stack creation is not sufficient for provisioning
a large number of nodes.  This allows the timeout for both undercloud and
overcloud stacks to be specified via environment varibles.  Both stacks
are created in Heat with a 6 hour timeout, and we rely entirely on
wait_for_stack_ready() to enforce the client-side timeout.  The default
remains 60 min.

Change-Id: I49ad0a6035392269d45b13a16fe0a78a61b38dfd