Ironic "lock already held" when powering off nodes
When we add a node devtests attempts to power it off the node. During the
time the power off command is run Ironic may be running a
periodic tasks to set the node power which takes an exclusive
lock. Hence making the power off command fail.

This patch will make the script run the ironic power-off 10 time
over 10 seconds to get the command to succeed.

Relate to: bug-1315224

Change-Id: I4d76bab7a3e07f6ff15f4e73440e298cc83eca12