Use sudo -E in install-dependencies
When following the install procedure in devtest.html,
any proxy information in the environment (eg, $http_proxy)
will not be seen by apt/yum/zypper being run inside
install-dependencies, as those commands are run inside sudo.

To preserve the environment values, this change uses "sudo -E"
so that these values should be accessible to the commands as
they're run.

Closes-Bug: #1324398
Change-Id: Ic757205f6c45d17e880cbedfe583b518a46bb107