Wait for os-collect-config to complete on the seed
Currently in devtest we wait for a node to become pingable and then
start configuring the seed details. However the node may not have
completed its configuration as os-collect-config may still be running
on the seed node, configuring low level services. This means we can
start configuring services too early and fail to correctly configure
the seed.

As the seed image carries local metadata and it is therefore not
under Heat control, we need to listen for and send a completion
signal. This allows us to detect when the seed has completed
configuration and the jumphost can safely continue.

Change-Id: Ibfd20bf632cdf3381615dad68ecf620685553d5f
Closes-Bug: #1329528