Wait for ssh demon to become active
With only a single SSH host key being checked, devtest attempts to connect to
the seed really quickly. Although devtest does wait for the node to become
pingable, this is not long enough for sshd to fully wake up on slow hosts.

This patch makes devtest wait until it can definitely ssh to the seed before
initialising Keystone.

Change-Id: Ibfd20bf632cdf3381615dad68ecf620685553d5f