Symink init-keystone into scripts, and use it
os-cloud-config now includes a init-keystone script, replacing the
one shipping in this repository. To make use of it, symlink it into
scripts so it's in the PATH and update the calls to use the new
options, as well as removing the old shell based scripts to avoid
confusing git and developers.

Change-Id: Ieb2df5fa1dcc51fd42ffa99c9124d2443efac3bf