Allow use of dib-utils
In order to use dib-run-parts from dib-utils instead of the one in
diskimage-builder, we need to pull down the dib-utils project and
add it to the path.  When not working from a git version (using a
package from either a distro or pypi), dib-run-parts would be
installed to the path, so this change does the same thing to allow
diskimage-builder to work consistently in all environments.

Required by Ia0a0df7784a14c49b5c47ac0b03e6c2602c84b3b

Change-Id: I0be1f876d0e4a7d38e0d5c6010a552a8ebb158a4