Symlink register-nodes into scripts, and use it
os-cloud-config now includes an register-nodes script, replacing
the one shipping in this repository. To make use of it, symlink
it into scripts so it's in the PATH and update the call in
setup-baremetal to use the new options, as well as removing the old
shell based script to avoid confusing git and developers.

Change-Id: I0ec217ee1d20ee02fc51c30d1d2d2529bb4b3d47