Split ssh keys into files per user.
Updates the way we maintain our SSH keys in TripleO to use
individual files per user.

This change is meant to support adding support for per user
ssh accounts. Right now SSH access to some of the TripleO
regions are provided directly via the /root account. Using
individual ssh accounts per user should allow some extra
security/auditability. This is a step in that direction...

For per user ssh access our convention will be that the IRC
nickname/username field should match the filename of the
ssh-keys file and will be used to provide a /home directory
and ssh access.

Change-Id: Id030c47ed8e246e7f26615640d5500148de1620e