Move user's group membership setup in separate script
This change moves the code needed to add the current user to the
libvirt group out of install-dependencies script and in a separate
script launched _at the end_ of devtest_setup.

Moving it at the end of devtest_setup is good for people running
the devtest_* scripts one by one as the process is not stopped
anymore in between the steps and there is no need to re-run
devtest_setup, after relogin one can continue with devtest_testenv.

Change-Id: Ica804c24bd2cbf70a7f2b9a34fa2e0495c85c456
Related-Bug: #1195046