Wait longer for BM node registration
At present the code waits for `nova hypervisor-stats` to show the
expected number of nodes are registered. However, at this stage the
nodes can be only partially registered. This means that when the
scheduler is asked for a host, it doesn't know EG that there is ram
available and the ram filter fails.

By also waiting for memory to be registered this patch should improve
the timing characteristics and avoid this failure.

Closes-Bug: 1339532
Change-Id: I2c08bb3dee44f002442fd5c764d5dffed47ee9ba