Tool to tie disconnected stacks together
This tool will deploy and update a set of disconnected stacks that need
to share limited information through parameters and outputs.

Much of this code is copy/pasted from python-heatclient's shell code as
there wasn't time to refactor python-heatclient to allow us to consume
the code directly.

An example of how a chunked master template is included as
chunk-heat-overcloud.yaml.

Change-Id: I99b4be2579e8a4f1f5836191d7c3343650cdb5a2