WIP: Add VLAN support to the seed.
This patch adds the ability to bring up a VLAN on the seed and undercloud,
which is needed to communicate with a VLAN if the undercloud public API will be
on a VLAN (and likewise overcloud). I'll be pushing this through to
the full stack before taking it out of WIP (in case the schema needs
to change etc). Current status is that we need to tweak the IP's we use and
document the impact of multi-homing; then we need to configure setup-neutron
for the undercloud, and then finally we get onto the overcloud.

Change-Id: I78c0fa3902c1e480b7ada0055332b14828c3c7f9