WIP: Add public VLAN support.
This patch adds the ability to bring up a VLAN on the seed and undercloud,
which is needed to communicate with a VLAN if the public APIs will be
on a VLAN. I'll be pushing this through to the full stack before taking it out
of WIP (in case the schema needs to change etc). Current status is that we need
to apply changes to the overcloud and document the impact of multi-homing on
routing / API use in docs.

Change-Id: I78c0fa3902c1e480b7ada0055332b14828c3c7f9
Co-Authored-By: Derek Higgins <derekh@redhat.com>
Closes-Bug: #1325114