Provide support for different gateway and seed ips
The gateway ip can be configured to be different than the seed ip.
If this is done however, there is no port configured as the gateway
which results in a non-functioning configuration. Someone may want
to move the gateway away from the seed to provide the ability to
shutdown the seed but continue to access the remainder of the cloud
environment.  If the gateway ip is different from the seed ip, a port
is created on brbm and assigned the gateway ip.

Change-Id: If35210827c3568d5c47fbae10db2b99083f0d3ab
Closes-Bug: 1337475