Move updating the BM networks to a new script
Currently, devtest_testenv will write the BM networks stanza into
the TE_DATAFILE, and since you can not run devtest_testenv inside
a test environment since it plays with VMs and so on, it makes it
impossible to change the network topology described by bm-networks.
To make it easy to update in a test environment, move the population
call to a separate script.

Change-Id: Iccc562ad16bf73118739cd4535abfac099834e3e