Make the tripleorc file set -u safe
We have had reports that when tripleorc is sourced by scripts which are
set -u (such as the "Devtest cloud initialization script" created by
Tuskar), execution is aborted because $ELEMENT_PATH is referenced before
it is set.

This change makes the first reference to $ELEMENT_PATH set -u safe - if
it's not set an empty string will be returned and the next line of code
will set it.

Change-Id: I0329fd4bdb6b5e360cf2464def6bace655827944
Closes-Bug: #1335105