Makes devtest_seed.sh respect local.json
The seed-stack-config element expects you to copy config.json
to local.json and add your seed settings. Devtest does this for
you but it ignores an existing local.json, which you may have
tweaked.

Update devtest_seed.sh so it will use local.json in preference
to config.json if it exists.

Change-Id: I0a3765c56352a4bca3d260b45774ce6ee098d806